requirements.txt
fastapi
uvicorn
openai
tavily-python
requests
beautifulsoup4
python-dotenv
sse-starlette
pydantic
